+++ channel/chat_channel.py
@@ -105,7 +105,7 @@ class ChatChannel(Channel):
 
         # 消息内容匹配过程，并处理content
         if ctype == ContextType.TEXT:
-            if first_in and "」\n- - - - - - -" in content:  # 初次匹配 过滤引用消息
+            if first_in and "」\n- - - - - - -" in content and not context.get("wechat_group_force_reply", False):  # 初次匹配 过滤引用消息
                 logger.debug(content)
                 logger.debug("[chat_channel]reference query skipped")
                 return None